Spy Games
Problem Statement
Hello lieutenant! It's with an emergency that our agency wishes to use the skills that you learned in the 'python intermediate' module. Your mission if you choose to accept it, would be to help us decipher a message that we received from our intel. We have multiple text files that need to be read and have certain operations performed to get our final message.

Good luck. The fate of humanity lies in your hands.

Outcomes
After completing this project, you will have a better understanding of solving logical problems using python. In this project, you will be applying the following concepts:

String operations
Conditional statement and loops
File I/O
Functions